refactor: move findClosestConfigFile
