+++ _deno-lock-config.ts
@@ -1,3 +1,29 @@
+import { dirname, join as pathjoin } from "@std/path";
+import { exists } from "@std/fs";
+
+/** find config file like a DOM's Element.closest() */
+export async function findClosestConfigFile(
+  startPath: string,
+  targetFiles: string[],
+): Promise<string | null> {
+  let currentPath = await Deno.realPath(startPath);
+
+  while (currentPath) {
+    for (const file of targetFiles) {
+      const filePath = pathjoin(currentPath, file);
+      if (await exists(filePath)) {
+        return filePath;
+      }
+    }
+
+    const parentPath = dirname(currentPath);
+    if (parentPath === currentPath) break;
+    currentPath = parentPath;
+  }
+
+  return null;
+}
+
 // deno.lock
 export interface LockConfig {
   version?: "4"; // we only support version 4
